Inspect assemblies in a temp path (bgo#751339)
The call to Assembly.GetTypes() requires that all the dependencies
of a certain assembly be in the same folder as where it is, otherwise
a ReflectionTypeLoadException could be thrown.

Plus, it's useful to PreLoad() the Assembly property in the ctor
of AssemblyItem, because this way we load all the main assemblies
before the dependencies of them (that is, we do a BFS search,
instead of a DFS), otherwise we could get a FileNotFoundException
when calling Assembly.Load().

This was happening when trying to bundle banshee, because for example
its Banshee.GStreamerSharp.dll assembly depended on Banshee.Services.dll
which was placed in a different folder.

The way to fix this is to copy all .NET assemblies in a first pass to
a temporary directory, where they can be inspected via reflection later.

Fixes https://bugzilla.gnome.org/show_bug.cgi?id=751339